User Interface
This is the User Interface for the music recommender. This currently holds the homepage where the user enters the artist's name. There are other pages that will be uploaded upon completion (artists profile with a list of songs and bio / social media links). The UI is written in HTML, styled with CSS, with functions from Javascript.